+++ R/read_package.R
@@ -1,11 +1,14 @@
 #' Read a Data Package descriptor file (`datapackage.json`)
 #'
-#' Reads information from a `datapackage.json` file, i.e. the
-#' [descriptor](https://specs.frictionlessdata.io/data-package/#descriptor) file
-#' that describes the Data Package metadata and its Data Resources.
+#' Reads information from a `datapackage.json` file, i.e. the [descriptor](
+#' https://specs.frictionlessdata.io/data-package/#descriptor) file that
+#' describes the Data Package metadata and its Data Resources.
+#'
+#' See `vignette("data-package")` to learn how this function implements the
+#' Data Package standard.
 #'
 #' @param file Path or URL to a `datapackage.json` file.
-#' @return Data Package object, see [create_package()].
+#' @return A Data Package object, see [create_package()].
 #' @family read functions
 #' @export
 #' @examples
@@ -19,9 +22,6 @@
 #' # Access the Data Package properties
 #' package$name
 #' package$created
-#'
-#' # List resources
-#' resources(package)
 read_package <- function(file = "datapackage.json") {
   # Read file
   if (!is.character(file)) {