+++ R/resources.R
@@ -8,9 +8,9 @@
 #' @export
 #' @examples
 #' # Load the example Data Package
-#' package <- example_package
+#' package <- example_package()
 #'
-#' # List resources
+#' # List the resources
 #' resources(package)
 resources <- function(package) {
   # Check package (and that all resource have a name)