+++ tests/testthat/test-add_resource.R
@@ -1,5 +1,5 @@
 test_that("add_resource() returns a valid Data Package", {
-  p <- example_package
+  p <- example_package()
   df <- data.frame("col_1" = c(1, 2), "col_2" = c("a", "b"))
   df_csv <- test_path("data/df.csv")
   schema <- create_schema(df)
@@ -21,7 +21,7 @@ test_that("add_resource() returns error on invalid Data Package", {
 
 test_that("add_resource() returns error when resource name contains invalid
            characters", {
-  p <- example_package
+  p <- example_package()
   df <- data.frame("col_1" = c(1, 2), "col_2" = c("a", "b"))
 
   # Invalid names
@@ -71,9 +71,24 @@ test_that("add_resource() returns error when resource name contains invalid
   expect_no_error(check_package(add_resource(p, "n.3-w_10", df)))
 })
 
+test_that("add_resource() returns error when replace is not a logical value", {
+  p <- example_package()
+  df <- data.frame("col_1" = c(1, 2), "col_2" = c("a", "b"))
+  expect_error(
+    add_resource(p, "new_resource", df, replace = "not_a_logical"),
+    class = "frictionless_error_replace_invalid"
+  )
+  expect_no_error(
+    check_package(add_resource(p, "new_resource", df, replace = TRUE))
+  )
+  expect_no_error(
+    check_package(add_resource(p, "new_resource", df, replace = FALSE))
+  )
+})
+
 test_that("add_resource() returns error when resource of that name already
-           exists", {
-  p <- example_package
+           exists (for default replace = FALSE)", {
+  p <- example_package()
   df <- data.frame("col_1" = c(1, 2), "col_2" = c("a", "b"))
   expect_error(
     add_resource(p, "deployments", df),
@@ -84,11 +99,16 @@ test_that("add_resource() returns error when resource of that name already
     regexp = "`package` already contains a resource named \"deployments\".",
     fixed = TRUE
   )
+  expect_error(
+    add_resource(p, "deployments", df),
+    regexp = "Use `replace = TRUE` to replace an existing resource.",
+    fixed = TRUE
+  )
 })
 
 test_that("add_resource() returns error when data is not data frame or
            character", {
-  p <- example_package
+  p <- example_package()
   expect_error(
     add_resource(p, "new", list()),
     class = "frictionless_error_data_type_invalid"
@@ -101,7 +121,7 @@ test_that("add_resource() returns error when data is not data frame or
 })
 
 test_that("add_resource() returns error on invalid or empty data frame", {
-  p <- example_package
+  p <- example_package()
   df <- data.frame("col_1" = c(1, 2), "col_2" = c("a", "b"))
   schema <- create_schema(df)
   expect_error(
@@ -118,7 +138,7 @@ test_that("add_resource() returns error on invalid or empty data frame", {
 
 test_that("add_resource() returns error if CSV file cannot be found", {
   skip_if_offline()
-  p <- example_package
+  p <- example_package()
   df_csv <- test_path("data/df.csv")
   schema <- create_schema(data.frame("col_1" = c(1, 2), "col_2" = c("a", "b")))
   expect_error(
@@ -142,13 +162,13 @@ test_that("add_resource() returns error if CSV file cannot be found", {
     class = "frictionless_error_path_not_found"
   )
   expect_error(
-    add_resource(p, "new", "http://example.com/no_such_file.csv"),
+    add_resource(p, "new", "https://example.com/no_such_file.csv"),
     class = "frictionless_error_url_not_found"
   )
 })
 
 test_that("add_resource() returns error on mismatching schema and data", {
-  p <- example_package
+  p <- example_package()
   df <- data.frame("col_1" = c(1, 2), "col_2" = c("a", "b"))
   df_csv <- test_path("data/df.csv")
   schema_invalid <- create_schema(df) # Not yet invalid
@@ -174,11 +194,11 @@ test_that("add_resource() returns error if ... arguments are unnamed", {
   df <- data.frame("col_1" = c(1, 2), "col_2" = c("a", "b"))
   schema <- create_schema(df)
   expect_error(
-    add_resource(p, "new", df, schema, delim = ",", "unnamed_value"),
+    add_resource(p, "new", df, schema, replace = FALSE, delim = ",", "unnamed"),
     class = "frictionless_error_argument_unnamed"
   )
   expect_error(
-    add_resource(p, "new", df, schema, delim = ",", "unnamed_value"),
+    add_resource(p, "new", df, schema, replace = FALSE, delim = ",", "unnamed"),
     "All arguments in `...` must be named.",
     fixed = TRUE
   )
@@ -219,7 +239,7 @@ test_that("add_resource() returns error if ... arguments are reserved", {
 })
 
 test_that("add_resource() adds resource", {
-  p <- example_package
+  p <- example_package()
   df <- data.frame("col_1" = c(1, 2), "col_2" = c("a", "b"))
   df_csv <- test_path("data/df.csv")
 
@@ -244,6 +264,16 @@ test_that("add_resource() adds resource", {
     resources(p),
     c("deployments", "observations", "media", "new_df", "new_csv")
   )
+})
+
+test_that("add_resource() can replace an existing resource", {
+  p <- example_package()
+  df <- data.frame("col_1" = c(1, 2), "col_2" = c("a", "b"))
+  expect_no_error(
+    add_resource(p, "deployments", df, replace = TRUE)
+  )
+  p_replaced <- add_resource(p, "deployments", df, replace = TRUE)
+  expect_equal(resources(p), resources(p_replaced))
 })
 
 test_that("add_resource() uses provided schema (list or path) or creates one", {
@@ -282,7 +312,7 @@ test_that("add_resource() uses provided schema (list or path) or creates one", {
 
 test_that("add_resource() can add resource from data frame, readable by
            read_resource()", {
-  p <- example_package
+  p <- example_package()
   df <- data.frame("col_1" = c(1, 2), "col_2" = c("a", "b"))
   p <- add_resource(p, "new", df)
   expect_identical(read_resource(p, "new"), dplyr::as_tibble(df))
@@ -291,7 +321,7 @@ test_that("add_resource() can add resource from data frame, readable by
 test_that("add_resource() can add resource from local, relative, absolute,
            remote or compressed CSV file, readable by read_resource()", {
   skip_if_offline()
-  p <- example_package
+  p <- example_package()
   schema <- get_schema(p, "deployments")
 
   # Local
@@ -316,8 +346,8 @@ test_that("add_resource() can add resource from local, relative, absolute,
 
   # Remote
   remote_path <- file.path(
-    "https://github.com/frictionlessdata/frictionless-r",
-    "raw/main/inst/extdata/deployments.csv"
+    "https://raw.githubusercontent.com/frictionlessdata/frictionless-r",
+    "main/inst/extdata/deployments.csv"
   )
   p <- add_resource(p, "remote", remote_path, schema)
   expect_identical(p$resources[[7]]$path, remote_path)